Refactor backup handling in backend implementations
- Introduced shared logic for beginning backups in `backend_shared_impl.h` and implemented it in `inproc_backend` and `standalone_backend`.
- Removed redundant backup handling code from `inproc_backend` and `standalone_backend`, delegating to the new shared method.
- Enhanced exception handling capabilities in the backup process.
- Added comprehensive unit tests for backup functionality, including version checks, epoch order validation, and exception scenarios.
- Updated test cases to reflect the new structure and ensure robust coverage of backup conditions.
