+++ test/limestone/grpc/backend_shared_impl_test.cpp
@@ -28,18 +28,45 @@
 #include "limestone/grpc/service/message_versions.h"
 #include "log_entry.h"
 
+namespace {
+
+using limestone::grpc::proto::BackupObjectType;
+
+// Structure representing backup file conditions for backup test
+struct backup_condition {
+    std::string path_pre_begin_backup;
+    std::string path_post_begin_backup;
+    std::string object_id;
+    std::string object_path;
+    BackupObjectType object_type;
+};
+
+static const std::vector<backup_condition> backup_conditions =
+    {{"blob/dir_00/00000000000000c8.blob", "blob/dir_00/00000000000000c8.blob", "", "", BackupObjectType::UNSPECIFIED},
+     {"compaction_catalog", "compaction_catalog", "compaction_catalog", "compaction_catalog", BackupObjectType::METADATA},
+     {"compaction_catalog.back", "compaction_catalog.back", "", "", BackupObjectType::UNSPECIFIED},
+     {"data/snapshot", "data/snapshot", "", "", BackupObjectType::UNSPECIFIED},
+     {"epoch", "epoch", "", "", BackupObjectType::UNSPECIFIED},
+     {"", "epoch.*.6", "epoch.*.6", "epoch.*.6", BackupObjectType::METADATA},
+     {"limestone-manifest.json", "limestone-manifest.json", "limestone-manifest.json", "limestone-manifest.json", BackupObjectType::METADATA},
+     {"pwal_0000.*.0", "pwal_0000.*.0", "pwal_0000.*.0", "pwal_0000.*.0", BackupObjectType::LOG},
+     {"pwal_0000.compacted", "pwal_0000.compacted", "pwal_0000.compacted", "pwal_0000.compacted", BackupObjectType::SNAPSHOT},
+     {"pwal_0001", "pwal_0001.*.0", "pwal_0001.*.0", "pwal_0001.*.0", BackupObjectType::LOG},
+     {"wal_history", "wal_history", "wal_history", "wal_history", BackupObjectType::METADATA}};
+}  // anonymous namespace
+
 namespace limestone::testing {
-
-
-
 
 using namespace limestone::grpc::backend;
 
 using limestone::api::log_entry;
 using limestone::grpc::backend::i_writer;
 using limestone::grpc::proto::BackupObject;
-using limestone::grpc::proto::BackupObjectType;
 using limestone::internal::wal_history;
+using limestone::grpc::proto::BeginBackupRequest;
+using limestone::grpc::proto::BeginBackupResponse;
+using limestone::grpc::service::begin_backup_message_version;
+using limestone::grpc::service::session_timeout_seconds;
 
 class backend_shared_impl_test : public backend_test_fixture {
 protected:
@@ -68,7 +95,68 @@ protected:
         bool fail_ = false;
     };
 
-protected:
+ protected:
+    template <typename Selector>
+    void assert_backup_file_conditions(Selector selector) {
+        namespace fs = boost::filesystem;
+        fs::path dir(get_location());
+
+        std::set<std::string> actual;
+        for (auto& entry : fs::recursive_directory_iterator(dir)) {
+            if (fs::is_regular_file(entry.status())) {
+                fs::path rel = fs::relative(entry.path(), dir);
+                std::string relstr = rel.generic_string();
+                actual.insert(relstr);
+            }
+        }
+
+
+        // Check for expected files
+        for (const auto& cond : backup_conditions) {
+            const std::string& pattern = selector(cond);
+            if (pattern.empty()) {
+                // If empty, the file must NOT exist
+                bool found = false;
+                for (const auto& act : actual) {
+                    if (act.empty()) {
+                        found = true;
+                        break;
+                    }
+                }
+                ASSERT_FALSE(found) << "Unexpected file found (should not exist): <empty pattern>";
+                continue;
+            }
+            bool found = false;
+            std::string regex_pat = wildcard_to_regex(pattern);
+            std::regex re(regex_pat);
+            for (const auto& act : actual) {
+                if (std::regex_match(act, re)) {
+                    found = true;
+                    break;
+                }
+            }
+            ASSERT_TRUE(found) << "Expected file pattern not found: " << pattern;
+        }
+
+        // Check for unexpected files
+        for (const auto& act : actual) {
+            bool matched = false;
+            for (const auto& cond : backup_conditions) {
+                const std::string& pattern = selector(cond);
+                if (pattern.empty()) {
+                    continue;
+                }
+                std::string regex_pat = wildcard_to_regex(pattern);
+                std::regex re(regex_pat);
+                if (std::regex_match(act, re)) {
+                    matched = true;
+                    break;
+                }
+            }
+            ASSERT_TRUE(matched) << "Unexpected file found: " << act;
+        }
+    }
+    
     void create_test_pwal_with_blogs() {
         gen_datastore();
         create_blob_file(*resolver_, 100);
@@ -1195,6 +1283,337 @@ TEST_F(backend_shared_impl_test, get_object_log_corrupted_file_returns_error_sta
     EXPECT_NE(std::string(status.error_message()).find("file is corrupted"), std::string::npos);
     EXPECT_TRUE(writer.responses.empty());
 }
+
+TEST_F(backend_shared_impl_test, begin_backup_version_boundary) {
+    gen_datastore();
+    prepare_backup_test_files();
+
+    backend_shared_impl backend(get_location(), 4);
+
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+
+    // version=0 (unsupported)
+    request.set_version(0);
+    request.set_begin_epoch(0);
+    request.set_end_epoch(0);
+    auto status = backend.begin_backup(*datastore_, &request, &response);
+    EXPECT_FALSE(status.ok());
+    EXPECT_EQ(status.error_code(), ::grpc::StatusCode::INVALID_ARGUMENT);
+
+    // version=1 (supported, but not implemented)
+    request.set_version(begin_backup_message_version);
+    status = run_with_epoch_switch([&]() { return backend.begin_backup(*datastore_, &request, &response); }, 7);
+    EXPECT_TRUE(status.ok());
+    
+    // version=2 (unsupported)
+    request.set_version(2);
+    status = backend.begin_backup(*datastore_, &request, &response);
+    EXPECT_FALSE(status.ok());
+    EXPECT_EQ(status.error_code(), ::grpc::StatusCode::INVALID_ARGUMENT);
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_overall) {
+    gen_datastore();
+    prepare_backup_test_files();
+    assert_backup_file_conditions([](const backup_condition& c) { return c.path_pre_begin_backup; });
+
+    // FLAGS_v = 50; // set VLOG level to 50
+
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    request.set_version(1);
+    request.set_begin_epoch(0);
+    request.set_end_epoch(0);
+    BeginBackupResponse response;
+    
+    // Call begin_backup via run_with_epoch_switch to synchronize with epoch switch and log rotation if needed
+    auto before = static_cast<int64_t>(std::time(nullptr));
+    auto status = run_with_epoch_switch([&]() { return backend.begin_backup(*datastore_, &request, &response); }, 7);
+    auto after = static_cast<int64_t>(std::time(nullptr));
+
+    // Check log_dir after begin_backup
+    assert_backup_file_conditions([](const backup_condition& c) { return c.path_post_begin_backup; });
+
+    // Check that session_id is a valid UUID (xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx)
+    const std::string& session_id = response.session_id();
+    std::regex uuid_regex(R"([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})");
+    EXPECT_TRUE(std::regex_match(session_id, uuid_regex)) << "session_id is not a valid UUID: " << session_id;
+
+    // expire_at should be in [before + session_timeout_seconds, after + session_timeout_seconds]
+    int64_t expire_at = response.expire_at();
+    EXPECT_GE(expire_at, before + session_timeout_seconds);
+    EXPECT_LE(expire_at, after + session_timeout_seconds);
+    EXPECT_EQ(response.start_epoch(), 0);
+    EXPECT_EQ(response.finish_epoch(), 0);
+
+    // --- Existing: verify contents of response.objects() ---
+    for (const auto& cond : backup_conditions) {
+        if (cond.object_id.empty() || cond.object_path.empty() || cond.object_type == BackupObjectType::UNSPECIFIED) {
+            continue; // Skip conditions that do not specify object details
+        }
+        bool found = false;
+        std::regex id_re(wildcard_to_regex(cond.object_id));
+        std::regex path_re(wildcard_to_regex(cond.object_path));
+        for (const auto& obj : response.objects()) {
+            if (std::regex_match(obj.object_id(), id_re) &&
+                std::regex_match(obj.path(), path_re) &&
+                obj.type() == cond.object_type) {
+                found = true;
+                break;
+            }
+        }
+        EXPECT_TRUE(found) << "BackupObject not found: id=" << cond.object_id << ", path=" << cond.object_path << ", type=" << cond.object_type;
+    }
+    // If there is an object in response.objects() that does not exist in backup_conditions, report an error
+    for (const auto& obj : response.objects()) {
+        bool found = false;
+        for (const auto& cond : backup_conditions) {
+            if (cond.object_id.empty() || cond.object_path.empty() || cond.object_type == BackupObjectType::UNSPECIFIED) {
+                continue;
+            }
+            std::regex id_re(wildcard_to_regex(cond.object_id));
+            std::regex path_re(wildcard_to_regex(cond.object_path));
+            if (std::regex_match(obj.object_id(), id_re) &&
+                std::regex_match(obj.path(), path_re) &&
+                obj.type() == cond.object_type) {
+                found = true;
+                break;
+            }
+        }
+        EXPECT_TRUE(found) << "Unexpected BackupObject found: id=" << obj.object_id() << ", path=" << obj.path() << ", type=" << obj.type();
+    }
+
+    const auto& session_store = backend.get_session_store();
+    auto session_opt = session_store.get_session(session_id);
+    ASSERT_TRUE(session_opt.has_value()) << "Session not found for session_id: " << session_id;
+    const auto& session = session_opt.value();
+
+    // Convert the backup_object map of the session to proto type and compare
+
+    std::vector<limestone::grpc::proto::BackupObject> session_objects;
+    for (auto it = session.begin(); it != session.end(); ++it) {
+        session_objects.push_back(it->second.to_proto());
+    }
+    // The number must match
+    EXPECT_EQ(session_objects.size(), static_cast<size_t>(response.objects_size()));
+    // Each element must match (order does not matter)
+    for (const auto& obj : session_objects) {
+        bool found = false;
+        for (const auto& resp_obj : response.objects()) {
+            if (obj.object_id() == resp_obj.object_id() &&
+                obj.path() == resp_obj.path() &&
+                obj.type() == resp_obj.type()) {
+                found = true;
+                break;
+            }
+        }
+        EXPECT_TRUE(found) << "Session backup_object not found in response: id=" << obj.object_id() << ", path=" << obj.path() << ", type=" << obj.type();
+    }
+    // And vice versa
+    for (const auto& resp_obj : response.objects()) {
+        bool found = false;
+        for (const auto& obj : session_objects) {
+            if (obj.object_id() == resp_obj.object_id() &&
+                obj.path() == resp_obj.path() &&
+                obj.type() == resp_obj.type()) {
+                found = true;
+                break;
+            }
+        }
+        EXPECT_TRUE(found) << "Response backup_object not found in session: id=" << resp_obj.object_id() << ", path=" << resp_obj.path() << ", type=" << resp_obj.type();
+    }
+
+    EXPECT_TRUE(status.ok());
+}
+
+// begin_epoch > end_epoch
+TEST_F(backend_shared_impl_test, begin_backup_epoch_order_ok) {
+    gen_datastore();
+    prepare_backup_test_files();
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+    request.set_version(begin_backup_message_version);
+    request.set_begin_epoch(3);
+    request.set_end_epoch(4);
+    auto status = run_with_epoch_switch([&]() { return backend.begin_backup(*datastore_, &request, &response); }, 7);
+    EXPECT_TRUE(status.ok());
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_epoch_order_equal_ng) {
+    gen_datastore();
+    prepare_backup_test_files();
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+    request.set_version(begin_backup_message_version);
+    request.set_begin_epoch(3);
+    request.set_end_epoch(3);
+    auto status = backend.begin_backup(*datastore_, &request, &response);
+    EXPECT_FALSE(status.ok());
+    EXPECT_EQ(status.error_code(), ::grpc::StatusCode::INVALID_ARGUMENT);
+    EXPECT_EQ(status.error_message(), "begin_epoch must be less than end_epoch: begin_epoch=3, end_epoch=3");
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_epoch_order_gt_ng) {
+    gen_datastore();
+    prepare_backup_test_files();
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+    request.set_version(begin_backup_message_version);
+    request.set_begin_epoch(4);
+    request.set_end_epoch(3);
+    auto status = backend.begin_backup(*datastore_, &request, &response);
+    EXPECT_FALSE(status.ok());
+    EXPECT_EQ(status.error_code(), ::grpc::StatusCode::INVALID_ARGUMENT);
+    EXPECT_EQ(status.error_message(), "begin_epoch must be less than end_epoch: begin_epoch=4, end_epoch=3");
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_begin_epoch_gt_snapshot_ok) {
+    gen_datastore();
+    prepare_backup_test_files();
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+    request.set_version(begin_backup_message_version);
+    EXPECT_EQ(snapshot_epoch_id_, 2);
+    request.set_begin_epoch(3);
+    request.set_end_epoch(4);
+    auto status = run_with_epoch_switch([&]() { return backend.begin_backup(*datastore_, &request, &response); }, 7);
+    EXPECT_TRUE(status.ok());
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_begin_epoch_eq_snapshot_ng) {
+    gen_datastore();
+    prepare_backup_test_files();
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+    request.set_version(begin_backup_message_version);
+    EXPECT_EQ(snapshot_epoch_id_, 2);
+    request.set_begin_epoch(2);
+    request.set_end_epoch(4);
+    auto status = backend.begin_backup(*datastore_, &request, &response);
+    EXPECT_FALSE(status.ok());
+    EXPECT_EQ(status.error_code(), ::grpc::StatusCode::INVALID_ARGUMENT);
+    EXPECT_EQ(status.error_message(), "begin_epoch must be strictly greater than the epoch id of the last snapshot: begin_epoch=2, snapshot_epoch_id=2");
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_begin_epoch_lt_snapshot_ng) {
+    gen_datastore();
+    prepare_backup_test_files();
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+    request.set_version(begin_backup_message_version);
+    EXPECT_EQ(snapshot_epoch_id_, 2);
+    request.set_begin_epoch(1);
+    request.set_end_epoch(4);
+    auto status = backend.begin_backup(*datastore_, &request, &response);
+    EXPECT_FALSE(status.ok());
+    EXPECT_EQ(status.error_code(), ::grpc::StatusCode::INVALID_ARGUMENT);
+    EXPECT_EQ(status.error_message(), "begin_epoch must be strictly greater than the epoch id of the last snapshot: begin_epoch=1, snapshot_epoch_id=2");
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_end_epoch_lt_current_ok) {
+    gen_datastore();
+    prepare_backup_test_files();
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+    request.set_version(begin_backup_message_version);
+    request.set_begin_epoch(3);
+    request.set_end_epoch(4); 
+    auto status = run_with_epoch_switch([&]() { return backend.begin_backup(*datastore_, &request, &response); }, 7);
+    EXPECT_TRUE(status.ok());
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_end_epoch_eq_current_ok) {
+    gen_datastore();
+    prepare_backup_test_files();
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+    request.set_version(begin_backup_message_version);
+    request.set_begin_epoch(3);
+    request.set_end_epoch(5);
+    auto status = run_with_epoch_switch([&]() { return backend.begin_backup(*datastore_, &request, &response); }, 7);
+    EXPECT_TRUE(status.ok());
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_end_epoch_gt_current_ng) {
+    gen_datastore();
+    prepare_backup_test_files();
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+    request.set_version(begin_backup_message_version);
+    request.set_begin_epoch(3);
+    request.set_end_epoch(6); 
+    auto status = backend.begin_backup(*datastore_, &request, &response);
+    EXPECT_FALSE(status.ok());
+    EXPECT_EQ(status.error_code(), ::grpc::StatusCode::INVALID_ARGUMENT);
+    EXPECT_EQ(status.error_message(), "end_epoch must be less than or equal to the current epoch id: end_epoch=6, current_epoch_id=5");
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_end_epoch_lt_boot_durable_epoch_ng) {
+    gen_datastore();
+    prepare_backup_test_files_without_compaction();
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+    request.set_version(begin_backup_message_version);
+    request.set_begin_epoch(1);
+    request.set_end_epoch(2); 
+    auto status = backend.begin_backup(*datastore_, &request, &response);
+    EXPECT_FALSE(status.ok());
+    EXPECT_EQ(status.error_code(), ::grpc::StatusCode::INVALID_ARGUMENT);
+    EXPECT_EQ(status.error_message(), "end_epoch must be strictly greater than the durable epoch id at boot time: end_epoch=2, boot_durable_epoch_id=3");
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_end_epoch_eq_boot_durable_epoch_ok) {
+    gen_datastore();
+    prepare_backup_test_files_without_compaction();
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+    request.set_version(begin_backup_message_version);
+    request.set_begin_epoch(1);
+    request.set_end_epoch(3); 
+    auto status = run_with_epoch_switch([&]() { return backend.begin_backup(*datastore_, &request, &response); }, 7);
+    EXPECT_TRUE(status.ok());
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_end_epoch_gt_boot_durable_epoch_ok) {
+    gen_datastore();
+    prepare_backup_test_files_without_compaction();
+    backend_shared_impl backend(get_location(), 4);
+    BeginBackupRequest request;
+    BeginBackupResponse response;
+    request.set_version(begin_backup_message_version);
+    request.set_begin_epoch(1);
+    request.set_end_epoch(4); 
+    auto status = run_with_epoch_switch([&]() { return backend.begin_backup(*datastore_, &request, &response); }, 7);
+    EXPECT_TRUE(status.ok());
+}
+
+TEST_F(backend_shared_impl_test, begin_backup_exception_handling) {
+    gen_datastore(call_ready_mode::call_ready_manual);
+    backend_shared_impl backend(get_location(), 4);
+    backend.set_exception_hook([]() { throw std::runtime_error("test exception"); });
+    BeginBackupRequest request;
+    request.set_version(begin_backup_message_version);
+    request.set_begin_epoch(1);
+    request.set_end_epoch(2);
+    BeginBackupResponse response;
+    auto status = backend.begin_backup(*datastore_, &request, &response);
+    EXPECT_EQ(status.error_code(), ::grpc::StatusCode::INTERNAL);
+    EXPECT_STREQ(status.error_message().c_str(), "test exception");
+}
+
+
 
 } // namespace limestone::testing
